+++ NEWS.md
@@ -1,10 +1,16 @@
+# Redatam R package version 1.1.0
+
+- Updated to Redatam engine 2024-12-22.
+- Added 'redatam_info()' function to get extended API information.
+- Added support to create Redatam databases: 'redatam_create()'
+
 # Redatam R package version 1.0.1
 
 -   Added package's source code URL (github repository)
 -   Added package's issues  URL (github Issues)
 -   Added entries to 'Articles': Crosstabs
 -   Updated function's documentation
--   Hided internal functions from documentation: redatam_init\_(), redatam_destroy\_(), redatam_internal_query(), redatam_internal_run()
+-   Hid internal functions from documentation: redatam_init\_(), redatam_destroy\_(), redatam_internal_query(), redatam_internal_run()
 
 # Redatam R package version 1.0.0